+++ collector/src/main/java/com/bitmovin/analytics/stateMachines/PlayerState.java
@@ -55,6 +55,7 @@ public enum PlayerState {
         @Override
         void onEnterState(PlayerStateMachine machine) {
             machine.enableRebufferHeartbeat();
+            machine.rebufferingTimeout.start();
         }
 
         @Override
@@ -64,6 +65,7 @@ public enum PlayerState {
                 long elapsedTimeOnEnter = machine.getElapsedTimeOnEnter();
                 listener.onRebuffering(elapsedTime - elapsedTimeOnEnter);
             }
+            machine.rebufferingTimeout.cancel();
         }
     },
     ERROR {